+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.1.25 / 2012-09-11
+
 * [feature] `shelly check` checks gems based on Cloudfile
 * [bug] Show 10 recent backups in `backup list` command, instead of the oldest